# api/schema/openapi.yaml
openapi: 3.1.0
info:
  title: CMCEN / RCMCE API
  version: 1.0.0
  description: >
    OpenAPI-style inventory for the Express routes mounted by server/server.js.
    Schemas are intentionally broad where handlers return Mongoose documents.
    The existing developer actor may save blank descriptive profile/editorial
    fields. Ordinary actors retain required checks but may retain authorized
    blanks on later edits. Credentials, identifiers, references, valid nonblank
    values, dates and publication requirements remain enforced; see API ROUTES.md.
    Every /api endpoint is rate limited by source IP and may return 429 with
    Retry-After and RateLimit-* headers. Sensitive endpoints have stricter limits.
    Every response includes X-Request-ID for support correlation. Server failures
    are logged and audited without retaining request bodies or query strings.
servers:
  - url: http://localhost:3000
    description: Local development (Eric)
  - url: http://localhost:3001
    description: Local development (Bray)
  - url: https://cmcen-staging.corebot.ca
    description: Staging
  - url: https://beta.cmcen-rcmce.ca
    description: Beta
security:
  - bearerAuth: []
...
paths:
...
  /api/admin/professional-awards/{awardId}/recipients/{recipientId}/news:
    post:
      tags: [Admin]
      summary: Create or reopen an optional news draft for a saved award recipient.
      description: >-
        Requires both award management and news permissions. Returns newsArticleId
        for a Content Workspace link. Reuses an existing linked article without
        changing its copy or publication state. New drafts contain suggested
        bilingual copy and the recipient photo, or the default crest.
        Concurrent requests reuse the same recipient-sourced draft.
      x-permission: canReviewAndPublish and canManageNews
      responses:
        "200":
          description: The existing or newly created article ID.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: The linked article is no longer available.
...
  /api/timers/active:
    get:
      tags: [Banners]
      summary: List active public banners.
      security: []
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [global, home]
            default: global
      responses:
        "200":
          description: Active banners.
...
  /api/admin/timers:
    description: A developer may save blank title and banner text. Timing and delivery validation is unchanged.
    get:
      tags: [Banners]
      summary: List all banners.
      x-permission: canManageTimers
      responses:
        "200":
          description: Banners.
    post:
      tags: [Banners]
      summary: Create banner.
      x-permission: canManageTimers
      requestBody:
        $ref: "#/components/requestBodies/Banner"
      responses:
        "201":
          description: Created banner.
  /api/admin/timers/{timerId}:
    description: Developer blank title and text saves retain timing, boolean and delivery validation.
    patch:
      tags: [Banners]
      summary: Update banner.
      x-permission: canManageTimers
      parameters:
        - $ref: "#/components/parameters/TimerId"
      requestBody:
        $ref: "#/components/requestBodies/Banner"
      responses:
        "200":
          description: Updated banner.
    delete:
      tags: [Banners]
      summary: Delete banner.
      x-permission: canManageTimers
      parameters:
        - $ref: "#/components/parameters/TimerId"
      responses:
        "200":
          description: Delete result.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TimerId:
      name: timerId
      in: path
      required: true
      schema: { type: string }
...
  requestBodies:
...
    Banner:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BannerInput"
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    LocalizedText:
      type: object
      properties:
        en: { type: string }
        fr: { type: string }
...
    BannerInput:
      type: object
      properties:
        title: { type: string, maxLength: 120 }
        text:
          $ref: "#/components/schemas/LocalizedText"
          description: Localized banner text. HTTP and HTTPS URLs in the text render as clickable links.
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        textColor:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        countdownAt:
          type: string
          format: date-time
          nullable: true
        placement:
          type: string
          enum: [global, home]
        screenPosition:
          type: string
          enum: [header, below-header]
        enabled:
          type: boolean
        dismissible:
          type: boolean
          default: true
          description: Allow dismissal for 24 hours per banner revision in the visitor's browser. Omitted PATCH values preserve the saved setting; legacy records default to true. Returned by banner GET endpoints.
        scrolling:
          type: boolean
          default: false
          description: Scroll the message with pause/resume controls; reduced-motion visitors see static text. Omitted PATCH values preserve the saved setting; legacy records default to false. Returned by banner GET endpoints.
        icon:
          type: string
          enum: [none, info, warning]
          default: warning
          description: Optional leading icon. Legacy records and API creates without this field use warning; omitted PATCH values preserve the saved choice. The new editor starts with none. Returned by banner GET endpoints.
        order:
          type: number
...